+++ src/fncs.cpp
@@ -5,6 +5,7 @@
 #include <algorithm>
 #include <cassert>
 #include <cctype>
+#include <cstdio>
 #include <cstdlib>
 #include <fstream>
 #include <functional>
@@ -48,6 +49,9 @@ static int simulation_id = 0;
 static int n_sims = 0;
 static fncs::time time_delta_multiplier = 0;
 static fncs::time time_delta = 0;
+static fncs::time time_peer = 0;
+static fncs::time time_current = 0;
+static fncs::time time_window = 0;
 static zsock_t *client = NULL;
 static map<string,string> cache;
 static vector<string> events;
@@ -66,7 +70,20 @@ static sub_string_t subs_string;
 
 
 #if defined(_WIN32)
-static void signal_handler_reset() { } /* no-op */
+static BOOL WINAPI
+s_handler_fn_shim (DWORD ctrltype)
+{
+    if (ctrltype == CTRL_C_EVENT) {
+        zctx_interrupted = 1;
+        zsys_interrupted = 1;
+    }
+    return FALSE;
+}
+static void signal_handler_reset()
+{
+    zsys_handler_set(NULL);
+    SetConsoleCtrlHandler(s_handler_fn_shim, TRUE);
+}
 #else
 #include <signal.h>
 static struct sigaction sigint_default;
@@ -114,7 +131,7 @@ void fncs::start_logging()
     const char *fncs_log_level = NULL;
     string simlog = simulation_name + ".log";
     bool log_file = false;
-    bool log_stdout = false;
+    bool log_stdout = true;
 
     /* name for fncs log file from environment */
     fncs_log_filename = getenv("FNCS_LOG_FILENAME");
@@ -131,7 +148,7 @@ void fncs::start_logging()
     /* whether to echo to stdout from environment */
     fncs_log_stdout = getenv("FNCS_LOG_STDOUT");
     if (!fncs_log_stdout) {
-        fncs_log_stdout = "no";
+        fncs_log_stdout = "yes";
     }
 
     /* whether to enable logging at all from environment */
@@ -151,17 +168,13 @@ void fncs::start_logging()
             || fncs_log_stdout[0] == 'n'
             || fncs_log_stdout[0] == 'F'
             || fncs_log_stdout[0] == 'f') {
-    }
-    else {
-        log_stdout = true;
+        log_stdout = false;
     }
 
-    if (fncs_log_file[0] == 'N'
-            || fncs_log_file[0] == 'n'
-            || fncs_log_file[0] == 'F'
-            || fncs_log_file[0] == 'f') {
-    }
-    else {
+    if (fncs_log_file[0] == 'Y'
+            || fncs_log_file[0] == 'y'
+            || fncs_log_file[0] == 'T'
+            || fncs_log_file[0] == 't') {
         log_file = true;
     }
 
@@ -187,6 +200,14 @@ void fncs::start_logging()
     }
 
     FNCSLog::ReportingLevel() = FNCSLog::FromString(fncs_log_level);
+}
+
+
+void fncs::replicate_logging(TLogLevel &level, FILE *& one, FILE *& two)
+{
+    level = FNCSLog::ReportingLevel();
+    one = Output2Tee::Stream1();
+    two = Output2Tee::Stream2();
 }
 
 
@@ -221,7 +242,7 @@ void fncs::initialize()
             YAML::Node doc;
             parser.GetNextDocument(doc);
             config = parse_config(doc);
-        } catch (YAML::ParserException &ex) {
+        } catch (YAML::ParserException &) {
             cerr << "could not open " << fncs_config_file << endl;
         }
     }
@@ -467,6 +488,17 @@ void fncs::initialize(Config config)
         LDEBUG2 << "key is " << key;
     }
 
+    /* next frame is peer time */
+    frame = zmsg_next(msg);
+    if (!frame) {
+        LERROR << "ACK message missing peer time";
+        die();
+        return;
+    }
+    long time_peer_long = atol(fncs::to_string(frame).c_str());
+    LDEBUG2 << "time_peer_long is " << time_peer_long;
+    time_peer = time_peer_long;
+
     /* last frame is second ACK */
     frame = zmsg_next(msg);
     if (!zframe_streq(frame, ACK)) {
@@ -477,6 +509,8 @@ void fncs::initialize(Config config)
     LDEBUG2 << "received second ACK";
     zmsg_destroy(&msg);
 
+    time_current = 0;
+    time_window = 0;
     is_initialized_ = true;
 }
 
@@ -487,34 +521,44 @@ bool fncs::is_initialized()
 }
 
 
-fncs::time fncs::time_request(fncs::time next)
+fncs::time fncs::time_request(fncs::time time_next)
 {
     LDEBUG4 << "fncs::time_request(fncs::time)";
 
     if (!is_initialized_) {
         LWARNING << "fncs is not initialized";
-        return next;
+        return time_next;
     }
 
-    fncs::time granted;
+    fncs::time time_granted;
+    fncs::time time_passed;
 
     /* send TIME_REQUEST */
-    LDEBUG2 << "sending TIME_REQUEST of " << next << " in sim units";
-    next *= time_delta_multiplier;
+    LDEBUG2 << "sending TIME_REQUEST of " << time_next << " in sim units";
+    time_next *= time_delta_multiplier;
 
-    if (next % time_delta != 0) {
+    if (time_next % time_delta != 0) {
         LERROR << "time request "
-            << next
-            << " is not a multiple of time delta ("
+            << time_next
+            << " ns is not a multiple of time delta ("
             << time_delta
-            << ")!";
+            << " ns)!";
         die();
-        return next;
+        return time_next;
     }
 
-    LDEBUG1 << "sending TIME_REQUEST of " << next << " nanoseconds";
-    zstr_sendm(client, fncs::TIME_REQUEST);
-    zstr_sendf(client, "%llu", next);
+    if (time_next < time_current) {
+        LERROR << "time request "
+            << time_next
+            << " ns is smaller than the current time ("
+            << time_current
+            << " ns)!";
+        die();
+        return time_next;
+    }
+
+    time_passed = time_next - time_current;
+    LDEBUG2 << "time advanced " << time_passed << " ns since last request";
 
     /* sending of the time request implies we are done with the cache
      * list, but the other cache remains as a last value cache */
@@ -524,6 +568,22 @@ fncs::time fncs::time_request(fncs::time next)
     for (clist_t::iterator it=cache_list.begin(); it!=cache_list.end(); ++it) {
         it->second.clear();
     }
+
+    if (time_passed < time_window) {
+        time_window -= time_passed;
+        LDEBUG1 << "there are " << time_window << " nanoseconds left in the window";
+        LDEBUG1 << "time_granted " << time_next << " nanoseonds";
+        time_current = time_next;
+        return time_next;
+    }
+    else {
+        LDEBUG1 << "time_window expired";
+        time_window = 0;
+    }
+
+    LDEBUG1 << "sending TIME_REQUEST of " << time_next << " nanoseconds";
+    zstr_sendm(client, fncs::TIME_REQUEST);
+    zstr_sendf(client, "%llu", time_next);
 
     /* receive TIME_REQUEST and perhaps other message types */
     zmq_pollitem_t items[] = { { zsock_resolve(client), 0, ZMQ_POLLIN, 0 } };
@@ -535,7 +595,7 @@ fncs::time fncs::time_request(fncs::time next)
         if (rc == -1) {
             LERROR << "client polling error: " << strerror(errno);
             die(); /* interrupted */
-            return next;
+            return time_next;
         }
 
         if (items[0].revents & ZMQ_POLLIN) {
@@ -548,7 +608,7 @@ fncs::time fncs::time_request(fncs::time next)
             if (!msg) {
                 LERROR << "null message received";
                 die();
-                return next;
+                return time_next;
             }
 
             /* first frame is message type identifier */
@@ -556,7 +616,7 @@ fncs::time fncs::time_request(fncs::time next)
             if (!frame) {
                 LERROR << "message missing type identifier";
                 die();
-                return next;
+                return time_next;
             }
             message_type = fncs::to_string(frame);
 
@@ -564,17 +624,17 @@ fncs::time fncs::time_request(fncs::time next)
             if (fncs::TIME_REQUEST == message_type) {
                 LDEBUG4 << "TIME_REQUEST received";
 
-                /* next frame is time */
+                /* time_next frame is time */
                 frame = zmsg_next(msg);
                 if (!frame) {
                     LERROR << "message missing time";
                     die();
-                    return next;
+                    return time_next;
                 }
                 /* convert time string to nanoseconds */
                 {
                     istringstream iss(fncs::to_string(frame));
-                    iss >> granted;
+                    iss >> time_granted;
                 }
 
                 /* destroy message early since a returned TIME_REQUEST
@@ -596,7 +656,7 @@ fncs::time fncs::time_request(fncs::time next)
                 if (!frame) {
                     LERROR << "message missing topic";
                     die();
-                    return next;
+                    return time_next;
                 }
                 topic = fncs::to_string(frame);
 
@@ -605,7 +665,7 @@ fncs::time fncs::time_request(fncs::time next)
                 if (!frame) {
                     LERROR << "message missing value";
                     die();
-                    return next;
+                    return time_next;
                 }
                 value = fncs::to_string(frame);
 
@@ -642,18 +702,29 @@ fncs::time fncs::time_request(fncs::time next)
             else {
                 LERROR << "unrecognized message type";
                 die();
-                return next;
+                return time_next;
             }
 
             zmsg_destroy(&msg);
         }
     }
 
-    LDEBUG1 << "granted " << granted << " nanoseonds";
+    LDEBUG1 << "time_granted " << time_granted << " nanoseonds";
+
+    time_current = time_granted;
+
+    /* the peers this sim interacts with have a larger 'tick' */
+    if (time_peer > time_delta) {
+        /* how much time is left before reaching the peers' time? */
+        time_window = time_peer - (time_current % time_peer);
+        LDEBUG1 << "new time_window of " << time_window << " nanoseconds";
+    }
+
     /* convert nanoseonds to sim's time unit */
-    granted = convert_broker_to_sim_time(granted);
-    LDEBUG2 << "granted " << granted << " in sim units";
-    return granted;
+    time_granted = convert_broker_to_sim_time(time_granted);
+    LDEBUG2 << "time_granted " << time_granted << " in sim units";
+
+    return time_granted;
 }
 
 
@@ -873,8 +944,8 @@ fncs::time fncs::time_unit_to_multiplier(const string &value)
 {
     LDEBUG4 << "fncs::time_unit_to_multiplier(string)";
 
-    fncs::time retval; 
-    fncs::time ignore; 
+    fncs::time retval = 0;
+    fncs::time ignore = 0;
     string unit;
     istringstream iss(value);
 
@@ -1006,7 +1077,7 @@ fncs::Config fncs::parse_config(const string &configuration)
             YAML::Node doc;
             parser.GetNextDocument(doc);
             config = parse_config(doc);
-        } catch (YAML::ParserException &ex) {
+        } catch (YAML::ParserException &) {
             cerr << "could not load YAML configuration string" << endl;
             cerr << "-- configuration was as follows --" << endl;
             cerr << configuration << endl;