+++ CHANGELOG.md
@@ -8,6 +8,16 @@ This project follows the [Gitflow Workflow model](https://www.atlassian.com/git/
 ## [Unreleased]
 The Unreleased section will be empty for tagged releases. Unreleased functionality appears in the develop branch.
 
+## [2.2.0] - 2016-06-02
+
+### Changed
+- Performance improvement when simulators have different time deltas.
+
+### Fixed
+- MATLAB hangs during exit. Caused by bad signal handler interaction.
+- fncs_broker logging was not working.
+- is_initialized was missing from C and Python APIs.
+
 ## [2.1.2] - 2016-03-29
 
 ### Fixed
@@ -44,7 +54,8 @@ This is the first production-ready release of the new FNCS library and
 broker.  This is a complete rewrite of the previous FNCS implementation
 and thus it gets the backwards-incompatible version 2.0 designation.
 
-[Unreleased]: https://github.com/FNCS/fncs/compare/v2.1.2...develop
+[Unreleased]: https://github.com/FNCS/fncs/compare/v2.2.0...develop
+[2.2.0]: https://github.com/FNCS/fncs/compare/v2.1.2...v2.2.0
 [2.1.2]: https://github.com/FNCS/fncs/compare/v2.1.1...v2.1.2
 [2.1.1]: https://github.com/FNCS/fncs/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/FNCS/fncs/compare/v2.0...v2.1.0